+++ wger/config/views/language_config.py
@@ -16,6 +16,7 @@
 
 import logging
 
+from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.utils.translation import ugettext as _
 from django.views.generic import UpdateView
@@ -27,7 +28,10 @@ from wger.utils.generic_views import WgerFormMixin
 logger = logging.getLogger(__name__)
 
 
-class LanguageConfigUpdateView(WgerFormMixin, UpdateView):
+class LanguageConfigUpdateView(WgerFormMixin,
+                               LoginRequiredMixin,
+                               PermissionRequiredMixin,
+                               UpdateView):
     '''
     Generic view to edit a language config
     '''
@@ -39,7 +43,7 @@ class LanguageConfigUpdateView(WgerFormMixin, UpdateView):
         '''
         Return to the language page
         '''
-        return reverse_lazy('config:language:view', kwargs={'pk': self.object.language_id})
+        return reverse_lazy('core:language:view', kwargs={'pk': self.object.language_id})
 
     def get_context_data(self, **kwargs):
         context = super(LanguageConfigUpdateView, self).get_context_data(**kwargs)