+++ wger/core/tests/test_sitemap.py
@@ -14,7 +14,7 @@
 
 from django.core.urlresolvers import reverse
 
-from wger.manager.tests.testcase import WorkoutManagerTestCase
+from wger.core.tests.base_testcase import WorkoutManagerTestCase
 
 
 class SitemapTestCase(WorkoutManagerTestCase):