+++ wger/core/views/languages.py
@@ -16,6 +16,7 @@
 
 import logging
 
+from django.contrib.auth.mixins import PermissionRequiredMixin, LoginRequiredMixin
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.utils.translation import ugettext as _
 from django.utils.translation import ugettext_lazy
@@ -31,15 +32,14 @@ from django.views.generic import (
 from wger.core.models import Language
 from wger.utils.generic_views import (
     WgerDeleteMixin,
-    WgerFormMixin,
-    WgerPermissionMixin
+    WgerFormMixin
 )
 
 
 logger = logging.getLogger(__name__)
 
 
-class LanguageListView(WgerPermissionMixin, ListView):
+class LanguageListView(LoginRequiredMixin, PermissionRequiredMixin, ListView):
     '''
     Show an overview of all languages
     '''
@@ -49,14 +49,14 @@ class LanguageListView(WgerPermissionMixin, ListView):
     permission_required = 'core.change_language'
 
 
-class LanguageDetailView(WgerPermissionMixin, DetailView):
+class LanguageDetailView(LoginRequiredMixin, PermissionRequiredMixin, DetailView):
     model = Language
     template_name = 'language/view.html'
     context_object_name = 'view_language'
     permission_required = 'core.change_language'
 
 
-class LanguageCreateView(WgerFormMixin, CreateView):
+class LanguageCreateView(WgerFormMixin, LoginRequiredMixin, PermissionRequiredMixin, CreateView):
     '''
     Generic view to add a new language
     '''
@@ -64,17 +64,18 @@ class LanguageCreateView(WgerFormMixin, CreateView):
     model = Language
     fields = '__all__'
     title = ugettext_lazy('Add')
-    form_action = reverse_lazy('config:language:add')
+    form_action = reverse_lazy('core:language:add')
     permission_required = 'core.add_language'
 
 
-class LanguageDeleteView(WgerDeleteMixin, DeleteView):
+class LanguageDeleteView(WgerDeleteMixin, LoginRequiredMixin, PermissionRequiredMixin, DeleteView):
     '''
     Generic view to delete an existing language
     '''
 
     model = Language
-    success_url = reverse_lazy('config:language:overview')
+    fields = '__all__'
+    success_url = reverse_lazy('core:language:overview')
     messages = ugettext_lazy('Successfully deleted')
     permission_required = 'core.delete_language'
 
@@ -85,19 +86,19 @@ class LanguageDeleteView(WgerDeleteMixin, DeleteView):
         context = super(LanguageDeleteView, self).get_context_data(**kwargs)
 
         context['title'] = _(u'Delete {0}?').format(self.object.full_name)
-        context['form_action'] = reverse('config:language:delete', kwargs={'pk': self.object.id})
+        context['form_action'] = reverse('core:language:delete', kwargs={'pk': self.object.id})
 
         return context
 
 
-class LanguageEditView(WgerFormMixin, UpdateView):
+class LanguageEditView(WgerFormMixin, LoginRequiredMixin, PermissionRequiredMixin, UpdateView):
     '''
     Generic view to update an existing language
     '''
 
     model = Language
     fields = '__all__'
-    form_action_urlname = 'config:language:edit'
+    form_action_urlname = 'core:language:edit'
     permission_required = 'core.change_language'
 
     def get_context_data(self, **kwargs):