+++ wger/manager/tests/test_email_reminder.py
@@ -14,12 +14,12 @@
 
 import datetime
 
-from django.core.management import call_command
-from django.core import mail
 from django.contrib.auth.models import User
-from wger.core.models import UserProfile
+from django.core import mail
+from django.core.management import call_command
 
-from wger.manager.tests.testcase import WorkoutManagerTestCase
+from wger.core.models import UserProfile
+from wger.core.tests.base_testcase import WorkoutManagerTestCase
 from wger.manager.models import Schedule
 from wger.manager.models import Workout