+++ wger/weight/tests/test_entry.py
@@ -14,15 +14,14 @@
 
 import datetime
 import decimal
+
 from django.core.urlresolvers import reverse
 
 from wger.core.tests import api_base_test
+from wger.core.tests.base_testcase import WorkoutManagerAddTestCase
+from wger.core.tests.base_testcase import WorkoutManagerEditTestCase, WorkoutManagerTestCase
 from wger.utils.constants import TWOPLACES
-
 from wger.weight.models import WeightEntry
-
-from wger.manager.tests.testcase import WorkoutManagerEditTestCase, WorkoutManagerTestCase
-from wger.manager.tests.testcase import WorkoutManagerAddTestCase
 
 
 class MealRepresentationTestCase(WorkoutManagerTestCase):